[WIP] Refactor launcher and add overlay support.
Refactored the launcher to improve app management and GUI updates. Introduced a new overlay module for modular GUI management and cleaner element handling. Added error handling for GUI element management and updated battery level calculation logic.

+++ lib/gsm/src/gsm.hpp
@@ -99,7 +99,7 @@ namespace GSM
     void acceptCall();
     void rejectCall();
 
-    int getBatteryLevel();
+    double getBatteryLevel();
     void getHour();
 
     int getNetworkStatus();
